refactor: extract migrations and seed logic from helpers.ts
Split lib/db/helpers.ts (719 lines) into focused modules:
- lib/db/migrations.ts: schema creation and ALTER TABLE migrations
- lib/db/seed.ts: exercise seeding and starter template repair
- lib/db/helpers.ts: reduced to DB init + query helpers (~75 lines)

Migration function split into 5 sub-functions to satisfy eslint
complexity and max-lines-per-function rules.
